requirements-vercel.txt
flask
flask-restx
aiohttp
asyncio
requests
python-multipart
aiodns
beautifulsoup4
python-dotenv
Werkzeug
urllib3
chardet
fake-useragent
Brotli 